requirements.txt
streamlit
pandas
plotly
bcrypt
python-dotenv
google-generativeai
